Remove "tee" driver
Due to concerns brought up on the mailing list I'm removing the "tee"
driver until it might have more use cases.

In the updated main(), if TTY1, USBKBD0, and FRAMEBUF all exist, then
TTY1 is opened on USBKBD0 and a shell is started using TTY1 as input and
FRAMEBUF as output.  In the Raspberry Pi's configuration there will
therefore be two independent shells, one associated with the serial port
and one associated with a USB keyboard and the framebuffer.